Flip IntegerChoice bounds and ChoiceValue::Integer to BigInt
IntegerChoice's min_value/max_value/shrink_towards and ChoiceValue::Integer
now hold the small-optimized BigInt instead of i128, so an integer choice can
represent arbitrary-precision integers while staying allocation-free for the
common i128-sized case. NodeSortKey's scalar magnitude widens to BigInt too.

draw_integer keeps its i128 convenience signature (for the many small-range
internal callers) and delegates to a new draw_integer_big; the boundary-biased
sampler and the f64-CDF distribution sampler generalize to BigInt bounds
(mirroring Hypothesis: random draws are capped near 2^127, with a uniform
fallback over the true range when the capped CDF window collapses).

The specialized integer shrink passes operate through an i128 "view"
(saturating bounds, skipping genuinely-huge values to the generic index
passes), preserving the exact existing shrink behaviour for every value that
fits i128. The database serialises integers as length-prefixed two's-complement
bytes, and the data tree keys on BigInt.

Schema interpretation still routes integer draws through the i128 path for now;
wiring arbitrary-precision bounds through the schema + HegelValue is the next
step.

+++ src/native/shrinker/index_passes.rs
@@ -152,7 +152,7 @@ impl<'a> Shrinker<'a> {
             for e in 0u32..11 {
                 let magnitude: i128 = 1i128 << e;
                 for &sign in &[1i128, -1] {
-                    let candidate_val = ChoiceValue::Integer(sign * magnitude);
+                    let candidate_val = ChoiceValue::Integer(BigInt::from(sign * magnitude));
                     if kind.validate(&candidate_val)
                         && candidate_val != node.value
                         && !candidates.contains(&candidate_val)
